factorisation du code dans main.py par des fonctions qui sont ensuite déplacées dans le script lexsub.py
