--- gold_fr
@@ -1,100 +1,0 @@
-avocat.n 1 :: homme de loi 2; orateur 1; accusateur 1; défenseur 1;
-avocat.n 2 :: fruit 1; fruit de l'avocatier 1;
-avocat.n 3 :: homme de loi 2; membre du barreau 1; défenseur 1; plaideur 1;
-avocat.n 4 :: défenseur 3; représentant légal 1;
-avocat.n 5 :: défenseur 2; juriste 1; représentant 1; représentant légal 1;
-avocat.n 6 :: fruit 1; fruit de l'avocatier 1;
-avocat.n 7 :: homme de loi 1; membre du barreau 1;
-avocat.n 8 :: homme de loi 1; membre du barreau 1; accusateur 1; défenseur 1;
-avocat.n 9 :: défenseur 3; champion 1; plaideur 1; promoteur 1;
-avocat.n 10 :: homme de loi 1; membre du barreau 1; accusateur 1; défenseur 1;
-baie.n 11 :: fruit 2; fruit sauvage 2; fruit rouge 1; airelle 1;
-baie.n 12 :: anse 3; crique 2; golfe 1;
-baie.n 13 :: fenêtre 3; ouverture 2;
-baie.n 14 :: fruit sauvage 2; fruit 1; airelle 1;
-baie.n 15 :: golfe 3; crique 1; anse 1;
-baie.n 16 :: fenêtre 3; ouverture 1; croisée 1;
-baie.n 17 :: golfe 3; anse 2; crique 1;
-baie.n 18 :: anse 3; crique 1; golfe 1;
-baie.n 19 :: fruit 2; fruit sauvage 2; fruit rouge 1; airelle 1;
-baie.n 20 :: anse 4; calanque 1; petite crique 1;
-carrière.n 21 :: vie professionnelle 3; activité professionnelle 1; parcours professionnel 1; vie active 1; activité 1;
-carrière.n 22 :: mine 2; exploitation 1; excavation 1; exploitation des pierres 1;
-carrière.n 23 :: profession 2; professionnelle 2; emploi 1; avancement 1; métier 1;
-carrière.n 24 :: oeuvre 2; parcours 1; parcours professionnel 1; travail 1;
-carrière.n 25 :: vie professionnelle 2; parcours professionnel 1; exercice 1; activité 1;
-carrière.n 26 :: réussite professionnelle 1;
-carrière.n 27 :: mine 2; exploitation 2; excavation 1;
-carrière.n 28 :: travail 1; profession 1; activité 1; métier 1; parcours 1; vie professionnelle 1; situation professionnelle 1; avenir professionnelle 1;
-carrière.n 29 :: activité 1; parcours 1; poste 1; emploi 1; profession 1;
-carrière.n 30 :: mine 2; excavation 1;
-feu.n 31 :: foyer 3; cheminée 2; âtre 1; bûcher 1;
-feu.n 32 :: phare 4; lumière 2;
-feu.n 33 :: température 3; puissance 1;
-feu.n 34 :: lumière 2; éclairage 1;
-feu.n 35 :: incendie 4;
-feu.n 36 :: signal lumineux 1; signal tricolore 1;
-feu.n 37 :: incendie 3; foyer 2; flammes 1;
-feu.n 38 :: flamme 2; passion 1; puissance 1; force 1; ardeur 1;
-feu.n 39 :: signal lumineux 1; signalisation routière 1;
-feu.n 40 :: incendie 4; flammes 2;
-glace.n 41 :: banquise 2; eau gelée 2;
-glace.n 42 :: miroir 4;
-glace.n 43 :: crème glacée 3;
-glace.n 44 :: crème glacée 3;
-glace.n 45 :: verglas 2; gel 1; sol gelé 1;
-glace.n 46 :: crème glacée 3;
-glace.n 47 :: vitres 3; miroir 1;
-glace.n 48 :: banquise 3; eau gelée 1;
-glace.n 49 :: sorbet 2; crème glacée 2;
-glace.n 50 :: gel 1; neige 1; givre 1; eau gelée 1;
-livre.n 51 :: oeuvre 2; ouvrage 2; bouquin 2; roman 2; recueil 1;
-livre.n 52 :: sterling 1; pound 1; livre sterling 1; unité monétaire 1;
-livre.n 53 :: oeuvre 3; ouvrage 2; roman 1; recueil 1;
-livre.n 54 :: ouvrage 3; oeuvre 2; texte 1; essai 1;
-livre.n 55 :: ouvrage 2; oeuvre 1; document 1; exemplaire 1; specimen 1;
-livre.n 56 :: tournois 1; unité monétaire 1;
-livre.n 57 :: roman 3; oeuvre 2; ouvrage 1; histoire 1; récit 1;
-livre.n 58 :: manuel 4; ouvrage 3; traité 1; document 1;
-livre.n 59 :: ouvrage 3; oeuvre 1; bouquin 1;
-livre.n 60 :: sterling 1; pound 1; livre sterling 1; unité monétaire 1;
-pièce.n 61 :: oeuvre 2; comédie 2; spectacle 1; représentation 1;
-pièce.n 62 :: document 4; papiers 2; justificatif 1;
-pièce.n 63 :: oeuvre 3; sculpture 1; photographies 1; peinture 1; tableau 1;
-pièce.n 64 :: salle 2; chambre 1;
-pièce.n 65 :: produit 1; article 1; marchandise 1; matériel 1; objet usiné 1;
-pièce.n 66 :: salle 4; espace 1;
-pièce.n 67 :: objet 2; accessoire 1; produit 1; élément 1; oeuvre 1; specimen 1;
-pièce.n 68 :: morceau 3; étude 2; sonte 1; concerto 1; oeuvre 1; composition 1;
-pièce.n 69 :: monnaie 3; monnaie métallique 1;
-pièce.n 70 :: morceau 3; bloc 1; objet usiné 1;
-reprise.n 71 :: occasion 3; moment 1; deux fois 1; par deux fois 1;
-reprise.n 72 :: retour 3; recommencement 2; redémarrage 1; relance 1;
-reprise.n 73 :: rachat 2; renflouage 1;
-reprise.n 74 :: répétition 1; adaptation 1; cover 1; inédit 1;
-reprise.n 75 :: répétition 3; poursuite 1;
-reprise.n 76 :: recommencement 2; rentrée 1; retour 1; commencement 1; rentrée des classes 1;
-reprise.n 77 :: rachat 2; restructuration 1;
-reprise.n 78 :: retour 1; redémarrage 1; rebond 1; relance 1;
-reprise.n 79 :: rentrée 2; retour 1; recommencement 1;
-reprise.n 80 :: relance 2; remontée 1; rebond 1;
-timbre.n 81 :: voix 3; intonation 2;
-timbre.n 82 :: sonorité 2; tonalité 2; ton 1; couleur 1;
-timbre.n 83 :: vignette 2; cachet 1;
-timbre.n 84 :: cachet 2; sceau 1;
-timbre.n 85 :: sceau 1; tampon 1; cachet 1; tonalité 1;
-timbre.n 86 :: sonorité 3; ton 1; intonation 1;
-timbre.n 87 :: sonorité 3; son 1; tonalité 1;
-timbre.n 88 :: sonorité 1; son 1; cachet 1; couleur 1;
-timbre.n 89 :: patch 3; cachet 1;
-timbre.n 90 :: vignette 2; cachet 1;
-voie.n 91 :: chemin 2; direction 1; route 1; trajectoire 1;
-voie.n 92 :: rue 3; route 2; accès 1;
-voie.n 93 :: passage 1; trace 1; rail 1;
-voie.n 94 :: parcours 1; chemin de fer 1; rail 1;
-voie.n 95 :: route 2; allée 1; chaussée 1; rue 1; passage 1; avenue 1; rocade 1;
-voie.n 96 :: comportement 1; méthode 1; approche 1; façon 1; manière 1;
-voie.n 97 :: moyen 3; vote 1; chemin 1; route 1; biais 1;
-voie.n 98 :: direction 1; piste 1; chemin 1; moyen 1;
-voie.n 99 :: direction 2; politique 1; chemin 1; orientation 1;
-voie.n 100 :: route 4; passage 1; chemin 1; avenue 1; autoroute 1; rocade 1;
